# docs/evolution/openapi-evolution.yaml
openapi: 3.0.3
info:
  title: Moonevue APIs - Current + Evolution
  version: 0.2.0
  description: |
    Contrato consolidado com:
    1) endpoints atuais implementados (gateway/auth)
    2) endpoints alvo para a evolução (v1) por fase
servers:
  - url: http://localhost:8080
    description: Gateway
  - url: http://localhost:8081
    description: Auth
security:
  - CookieAuth: []
...
paths:
  /clients:
    get:
      tags: [Current Clients]
      summary: List clients (current)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Clients page
    post:
      tags: [Current Clients]
      summary: Create client (current)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUpsertRequest'
      responses:
        '201':
          description: Created
  /clients/{clientId}:
    get:
      tags: [Current Clients]
      summary: Get client by id (current)
      parameters:
        - $ref: '#/components/parameters/ClientIdLong'
      responses:
        '200':
          description: Client
    put:
      tags: [Current Clients]
      summary: Update client (current)
      parameters:
        - $ref: '#/components/parameters/ClientIdLong'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUpsertRequest'
      responses:
        '200':
          description: Updated
  /clients/{clientId}/transactions:
    get:
      tags: [Current Clients, Current Payments]
      summary: List transactions by client (current)
      parameters:
        - $ref: '#/components/parameters/ClientIdLong'
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Transactions page
...
  /checkout/{token}:
    get:
      tags: [Current Checkout]
      security: []
      summary: Get checkout info (current)
      parameters:
        - $ref: '#/components/parameters/CheckoutTokenUuid'
      responses:
        '200':
          description: Checkout info
  /checkout/{token}/status:
    get:
      tags: [Current Checkout]
      security: []
      summary: Get checkout status (current)
      parameters:
        - $ref: '#/components/parameters/CheckoutTokenUuid'
      responses:
        '200':
          description: Checkout status
  /checkout/{token}/client-lookup:
    get:
      tags: [Current Checkout]
      security: []
      summary: Lookup client by document (current)
      parameters:
        - $ref: '#/components/parameters/CheckoutTokenUuid'
        - in: query
          name: document
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Lookup result
  /checkout/{token}/identify:
    post:
      tags: [Current Checkout]
      security: []
      summary: Identify checkout payer (current)
      parameters:
        - $ref: '#/components/parameters/CheckoutTokenUuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutIdentifyRequest'
      responses:
        '200':
          description: Identified
  /checkout/{token}/pay:
    post:
      tags: [Current Checkout]
      security: []
      summary: Pay checkout transaction (current)
      parameters:
        - $ref: '#/components/parameters/CheckoutTokenUuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutPayRequest'
      responses:
        '200':
          description: Payment requested
...
  /auth/employees:
    post:
      tags: [Current Employees]
      summary: Create employee user (current)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeRegisterRequest'
      responses:
        '201':
          description: Employee created
...
  /v1/checkout/{token}/identify:
    post:
      tags: [Evolution Checkout]
      security: []
      summary: Identify/create customer in checkout (target)
      parameters:
        - $ref: '#/components/parameters/CheckoutTokenText'
      responses:
        '200':
          description: Identified
  /v1/checkout/{token}/pay:
    post:
      tags: [Evolution Checkout]
      security: []
      summary: Execute checkout payment (target)
      parameters:
        - $ref: '#/components/parameters/CheckoutTokenText'
      responses:
        '200':
          description: Initiated
  /v1/checkout/{token}/status:
    get:
      tags: [Evolution Checkout]
      security: []
      summary: Poll checkout status (target)
      parameters:
        - $ref: '#/components/parameters/CheckoutTokenText'
      responses:
        '200':
          description: Status
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: sid
  parameters:
    ClientIdLong:
      in: path
      name: clientId
      required: true
      schema: { type: integer, format: int64 }
...
    CheckoutTokenUuid:
      in: path
      name: token
      required: true
      schema: { type: string, format: uuid }
    CheckoutTokenText:
      in: path
      name: token
      required: true
      schema: { type: string }
  schemas:
    ClientUpsertRequest:
      type: object
      required: [name, cpfCnpj, email]
      properties:
        name: { type: string }
        cpfCnpj: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
    EmployeeRegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
...
    CheckoutIdentifyRequest:
      type: object
      required: [document]
      properties:
        document: { type: string }
    CheckoutPayRequest:
      type: object
      required: [instrument, payerName, payerDocument]
      properties:
        instrument: { type: string }
        payerName: { type: string }
        payerDocument: { type: string }
        payerEmail: { type: string, format: email, nullable: true }
        payerPhone: { type: string, nullable: true }
        pixKey: { type: string, nullable: true }
...