# services/backend/gateway/src/main/java/com/moonevue/gateway/controller/CheckoutController.java
package com.moonevue.gateway.controller;

import com.moonevue.gateway.dto.CheckoutInfoDTO;
import com.moonevue.gateway.dto.CheckoutClientLookupDTO;
import com.moonevue.gateway.dto.CheckoutIdentifyRequest;
import com.moonevue.gateway.dto.CheckoutPayRequest;
import jakarta.validation.Valid;
import com.moonevue.gateway.service.CheckoutService;
import com.moonevue.gateway.service.IdempotencyService;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.Map;
import java.util.UUID;

@RestController
@RequestMapping("/checkout")
public class CheckoutController {

    private static final Logger log = LoggerFactory.getLogger(CheckoutController.class);

    private final CheckoutService checkoutService;
    private final IdempotencyService idempotencyService;

    public CheckoutController(CheckoutService checkoutService,
                              IdempotencyService idempotencyService) {
        this.checkoutService = checkoutService;
        this.idempotencyService = idempotencyService;
    }

    @GetMapping("/{token}")
    public ResponseEntity<?> getInfo(@PathVariable("token") UUID token) {
        try {
            CheckoutInfoDTO info = checkoutService.getInfo(token);
            return ResponseEntity.ok(info);
        } catch (IllegalArgumentException e) {
            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Map.of("error", e.getMessage()));
        } catch (IllegalStateException e) {
            return ResponseEntity.status(HttpStatus.GONE).body(Map.of("error", e.getMessage()));
        } catch (Exception e) {
            log.error("[CheckoutController] Erro ao buscar checkout token={}: {}", token, e.getMessage(), e);
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(Map.of("error", "Erro ao carregar link"));
        }
    }

    @GetMapping("/{token}/status")
    public ResponseEntity<?> getStatus(@PathVariable("token") UUID token) {
        try {
            CheckoutInfoDTO info = checkoutService.getStatus(token);
            return ResponseEntity.ok(info);
        } catch (IllegalArgumentException e) {
            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Map.of("error", e.getMessage()));
        } catch (IllegalStateException e) {
            return ResponseEntity.status(HttpStatus.GONE).body(Map.of("error", e.getMessage()));
        } catch (Exception e) {
            log.error("[CheckoutController] Erro ao buscar status token={}: {}", token, e.getMessage(), e);
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(Map.of("error", "Erro ao consultar status"));
        }
    }

    @GetMapping("/{token}/client-lookup")
    public ResponseEntity<?> lookupClient(
            @PathVariable("token") UUID token,
            @RequestParam("document") String document
    ) {
        try {
            CheckoutClientLookupDTO result = checkoutService.lookupClient(token, document);
            return ResponseEntity.ok(result);
        } catch (IllegalArgumentException e) {
            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Map.of("error", e.getMessage()));
        } catch (IllegalStateException e) {
            return ResponseEntity.status(HttpStatus.GONE).body(Map.of("error", e.getMessage()));
        } catch (Exception e) {
            log.error("[CheckoutController] Erro ao buscar cliente token={}: {}", token, e.getMessage(), e);
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(Map.of("error", "Erro ao consultar cliente"));
        }
    }

    @PostMapping("/{token}/identify")
    public ResponseEntity<?> identify(@PathVariable("token") UUID token,
                                      @Valid @RequestBody CheckoutIdentifyRequest request) {
        try {
            CheckoutInfoDTO info = checkoutService.identify(token, request.document());
            return ResponseEntity.ok(info);
        } catch (IllegalArgumentException e) {
            return ResponseEntity.badRequest().body(Map.of("error", e.getMessage()));
        } catch (IllegalStateException e) {
            return ResponseEntity.status(HttpStatus.GONE).body(Map.of("error", e.getMessage()));
        } catch (Exception e) {
            log.error("[CheckoutController] Erro ao identificar cliente token={}: {}", token, e.getMessage(), e);
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(Map.of("error", "Erro ao validar identidade"));
        }
    }

    @PostMapping("/{token}/pay")
    public ResponseEntity<?> pay(@PathVariable("token") UUID token,
                                 @RequestHeader(value = "Idempotency-Key", required = false) String idempotencyKey,
                                 @RequestBody CheckoutPayRequest request) {
        String scope = "checkout_pay_" + token;
        String requestHash = idempotencyService.hashRequest(request == null ? Map.of() : request);
        if (idempotencyKey != null && !idempotencyKey.isBlank()) {
            var existing = idempotencyService.findByScopeAndKey(scope, idempotencyKey);
            if (existing.isPresent() && existing.get().responseStatus() != null && existing.get().responseBody() != null) {
                if (!existing.get().requestHash().equals(requestHash)) {
                    return ResponseEntity.status(HttpStatus.CONFLICT)
                            .body(Map.of("error", "Idempotency-Key reutilizada com payload diferente"));
                }
                CheckoutInfoDTO cached = idempotencyService.parseResponseBody(existing.get().responseBody(), CheckoutInfoDTO.class);
                return ResponseEntity.status(existing.get().responseStatus()).body(cached);
            }
            idempotencyService.reserve(scope, idempotencyKey, requestHash);
        }

        try {
            CheckoutInfoDTO result = checkoutService.pay(token, request);
            if (idempotencyKey != null && !idempotencyKey.isBlank()) {
                idempotencyService.storeResponse(scope, idempotencyKey, HttpStatus.OK.value(), result);
            }
            return ResponseEntity.ok(result);
        } catch (IllegalArgumentException e) {
            return ResponseEntity.badRequest().body(Map.of("error", e.getMessage()));
        } catch (IllegalStateException e) {
            return ResponseEntity.status(HttpStatus.GONE).body(Map.of("error", e.getMessage()));
        } catch (Exception e) {
            log.error("[CheckoutController] Falha ao pagar checkout token={}: {}", token, e.getMessage(), e);
            return ResponseEntity.status(HttpStatus.BAD_GATEWAY)
                    .body(Map.of("error", "Falha ao processar pagamento", "detail", e.getMessage()));
        }
    }
}